+++ travel-packages/infraestructura/src/main/java/com/ceiba/paquete/adaptador/dao/MapeoPaquete.java
@@ -12,7 +12,7 @@ public class MapeoPaquete implements RowMapper<DtoPaquete>, MapperResult {
     public DtoPaquete mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         return new DtoPaquete(
                 resultSet.getLong("id"),
-                resultSet.getLong("precio"),
+                resultSet.getDouble("precio"),
                 resultSet.getString("estado"),
                 resultSet.getString("ciudad"),
                 resultSet.getString("hotel"),
